I think now everything works fine ( at least for .so files and gmake ). If someone else could check it would be great.
Jni is built if --enable-jni configure option is used, if something change it'll
be recompiled ( without recompiling everything else ). I'm not sure how makdepend
could be incoprorated ( probably making it write in Makefile.in ).

I'm also not sure why Makefile.am in 1.3 and Makefile.in in 2.0.
Another question - do we need Makefile in common ?

In any case, I believe the only way to get a decent build system is to replace
this mess with ant. The options should be specified in build.properties, not in
command line, and the complexity should be hidden in ant tasks, not in makefiles
and shell scripts.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@295882 13f79535-47bb-0310-9956-ffa450edef68