The big commit - plugin in the new socket code.
All is #ifdef-ed by now - the old code is compiled by default, unmodified
( except some ifs that I reversed to be able to read and minor indentations).

Probably this will remain this way until we are comfortable with the new
abstraction - then we can remove the old code and the ifdefs. So far I can run it without
problems.


One big issue is the code that recycles the connection - I'm very comfused, it
seems that besides recycling the endpoint we have some code that is doing
some tricks with the sockets. From the comments it seems to be dealing with
lb - but the lb worker has it's own mechanism to get another endpoint
if one fails.

I've also got a core on high load ( but very hard to reproduce ) and it happens
in the reuse_connection - I added a check, but not sure why the worker would be null
in the endpoint. What's more important is finding if we need this extra attempt to reuse
( or recover from endpoint errors ), closing the socket and reopening again could
be a better solution ( if tomcat was restarted, all sockets will be invalid, so
reuse_connection couldn't help in any way ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@295982 13f79535-47bb-0310-9956-ffa450edef68