Initial implementation for the apr pool.
This will allow use of native pools if apache is used ( or apr ). Non-apr
code will work as before, using jk pools.

Note that for Apache1.3 we also have the option to use apache1.3 pools -
but probably it's not worth the effort ( it's easier to use jk_pool and
better to use apr pools )


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@296091 13f79535-47bb-0310-9956-ffa450edef68