- Added pools wherever they're needed.   Factory now takes a pool - that allows us to get rid of 1/2 of mallocs and be sure the data will be cleaned.
- Pass pool to methods that could be called at run time and may need
to create temp data ( we don't want to allocate temp data in the object's pool,
but in the caller pool or a temp pool - which is short-lived )

- Get rid of all buf[]s. Now all pools are created using the pool hierarchy -
i.e. from apr_pools or malloc. Since we recycle, there's no need to play all
the stack-allocation tricks.

- moved the jk_pool-specific private data to the impl. The common implementation
will be used only if APR ( or other server-specific native pools ) are not available.

- removed the per/context fileds from the per/uri structure


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@296102 13f79535-47bb-0310-9956-ffa450edef68