As previously discussed, add jk_env param to all methods.
This is for now just a 'cosmetic' change - all functions will now follow
the same pattern ( previously most but not all had access to logger, etc).
The actual implementation of env is not completed, but should work the same
as it did before - it's just one level of indirection to access the logger.
We still use a single shared logger, etc.

In future we may improve this to give each thread a env instance ( that will
include a per thread logger and per thread temp pool )
that will allow us to implement error handling and
get rid of the last 3 mallocs ( and use pools exclusively ).

This should be my last 'major' change in jk2 refactoring.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@296166 13f79535-47bb-0310-9956-ffa450edef68