Few more changes.
Consolidate all build.properties in the top dir ( so user has to change
 a single file ).

Add 'install' option that will copy the files in tomcat ( for development it's
faster to use symlinks ).

Added code to detect 4.0/4.1 and build with 4.1 if available.

( now the compile fails if you have 4.1 detected - comment it
in build.properties ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@296190 13f79535-47bb-0310-9956-ffa450edef68