Some more changes in mod_jk configuration, to provide a more 'natural' style for apache users.
As previously, worker.properties is no longer needed ( but can be used ),
all settings can be done in httpd.conf ( or generated include file )
using "JkSet name value"

Adding 'mounts' is done using <Locations> - and the mapping is done
by mod_core only once. I changed the command ( it was added few weeks
ago ) to "JkWebapp worker ajp13" ( other options besides 'worker' will
be available ).

The JkWebapp should be specified for <Locations> coresponding the
the root of the webapp. Autoconf will get the other mappings, or
generated config.

I added a different "JkServlet" command to be used for all servlet
mappings inside a webapp, it'll eventually allow jk to do the
corect mapping and pass the processed info to tomcat, to avoid duplicating
it.

Still to do - remove the need to declare WEB-INF/META-INF constraints
( we can code this instead of requiring the user to do it ).

( old config should still work )


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@296203 13f79535-47bb-0310-9956-ffa450edef68