Moved that strange is_recoverable_error in request instead of passing it as argument to all methods.
The real meaning ( not easy to discover :-) is if the request can be safely
sent to a fallback worker. It is used by lb, but all workers doing active
processing must set it ( or leave a default FALSE ).

Init() method will now use the worker, not the endpoint. It is perfectly
possible to recycle requests or preinit them or anything else - they don't
depend on a particular connection.

( I also added flush, probably not needed but I have a feeling it's missing )


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk/jk@296314 13f79535-47bb-0310-9956-ffa450edef68